qr_app/requirements.txt
qrcode 
pillow 
flask